% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieCuda.R
\name{.callbinary2}
\alias{.callbinary2}
\title{Make system call for binaries}
\usage{
.callbinary2(bin1, args1, path = NULL)
}
\arguments{
\item{bin1}{\code{Character}. The binary used for the system call.}

\item{args1}{\code{Character}. The arguments to pass to the binary.}

\item{path}{\code{Character} Optional: If passed to function, returns
the path.}
}
\value{
The output of the system call or the path provided.
}
\description{
Function that makes a system call for the RbowtieCuda binaries.
 Note it is not intended to be used outside of the package.
}
\author{
Franck RICHARD
}
