% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieCuda.R
\name{nvBowtie_usage}
\alias{nvBowtie_usage}
\title{Print available arguments that can be passed to nvBowtie()}
\usage{
nvBowtie_usage()
}
\value{
Information about available arguments that can be passed to nvBowtie().
}
\description{
Calling nvBowtie_usage() prints the available arguments that can
be passed to the ... argument 'options' of the nvBowtie() function of the package.
}
\examples{
nvBowtie_usage()
}
\references{
Langmead, B., & Salzberg, S. L. (2012). Fast gapped-read
alignment with Bowtie 2. Nature methods, 9(4), 357-359.
}
\author{
Franck RICHARD
}
