% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieCuda.R
\name{.callbinary2}
\alias{.callbinary2}
\title{Make system call for binaries}
\usage{
.callbinary2(bin1, args1, path = NULL, code = NULL)
}
\arguments{
\item{bin1}{\code{Character}. The binary used for the system call.}

\item{args1}{\code{Character}. The arguments to pass to the binary.}

\item{path}{\code{Character} Optional: If passed to function, returns
the path.}
\item{code}{\code{Character} Optional: If passed to function, returns
the code.}
}
\value{
An invisible \code{Integer} of call status.
The value is 0 when there is not any mistakes.
}
\description{
Function that makes a system call for the RbowtieCuda binaries.
 Note that it is not designed to be used outside the package.
}
\examples{
.callbinary2(bin1 = "nvBowtie", args1 = "--version")
}
\author{
Franck RICHARD
}
