% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieCuda.R
\name{nvbio_tests}
\alias{nvbio_tests}
\title{Print unit tests}
\usage{
nvbio_tests()
}
\value{
The value is 0 when there is not any mistakes.
}
\description{
Calling nvbio_tests() performs alignment tests
}
\examples{
nvbio_tests()
}
\author{
Franck RICHARD
}
