% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtieCuda.R
\name{nvBowtie_version}
\alias{nvBowtie_version}
\title{Print version information}
\usage{
nvBowtie_version()
}
\value{
An invisible \code{Integer} of call status.
The value is 0 when there is not selected
}
\description{
Calling nvBowtie_version() prints the version information of 
the RBowtieCuda package used.
}
\examples{
nvBowtie_version()
}
\author{
Franck RICHARD
}
